Update package.json
Solve the error Message "The library may need to update its package.json or typings.ts" when trying to import webgl-fluid-enhanced on React TS environment. 

Type is located at the root of the package.json but need to be within "." object. 

Please let me know if some errors occurs with JS